requirements.txt
flask
openai
python-dotenv
jinja2



